requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-rest-framework==0.1.0
django-utils==0.0.2
djangorestframework==3.15.2
environ==1.0
Markdown==3.7
mysqlclient==2.2.4
packaging==24.1
pycparser==2.22
sqlparse==0.5.1
tzdata==2024.1
utils==1.0.2
